// frontend/packages/workflow/playground/src/node-registries/mirap-mmsi-set/types.ts
/*
 * Copyright 2025 coze-dev Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { InputValueVO, ValueExpression } from '@coze-workflow/base';
import { type ViewVariableType } from '@coze-workflow/base';

export enum MMSISetOperation {
  Intersection = 'intersection',
  Union = 'union',
  Difference = 'difference',
}

export interface SelectedOutputGroup {
  inputName: string;
  fields: string[];
}

export interface OutputFieldMeta {
  name: string;
  type: ViewVariableType;
}

export interface OutputFieldGroup {
  inputName: string;
  title: string;
  fields: OutputFieldMeta[];
}

export interface FormData {
  inputs: {
    inputParameters: InputValueVO[];
    selectedOutputGroups: SelectedOutputGroup[];
    mainInputName?: string;
    [key: string]: unknown;
  };
  outputs: Array<{
    key?: string;
    name: string;
    type: ViewVariableType;
    children?: OutputFieldMeta[];
  }>;
}

export interface InputParameterDTO {
  name?: string;
  input?: ValueExpression;
}